pyproject.toml
[project]
name = "deutsch-tg-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "anthropic~=0.74.0",
    "pydantic-settings~=2.12.0",
    "python-telegram-bot~=22.5",
    "rich~=14.2.0",
    "typer~=0.20.0",
]

[dependency-groups]
dev = [
    "mypy==1.18.2",
    "pre-commit~=4.4.0",
    "ruff~=0.14.0",
]

[tool.ruff]
line-length = 100
extend-exclude = []

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
extend-select = ["E501", "RUF100", "I"]

[tool.ruff.lint.pycodestyle]
max-line-length = 120

[tool.mypy]
packages = ["deutsch_tg_bot"]
python_version = "3.14"
# mypy_path = "mypy_stubs"
strict = true
warn_return_any = true
warn_unused_configs = true
enable_error_code = [
    "truthy-bool",
]
